Update FragmentExamplesActivity
- Fix behavior when "Don't keep activities" is enabled
- Create `FragmentExamplesViewModel`
- Move fragment to upper level class
