+++ example/res/layout/fragments_example_activity.xml
@@ -1,13 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<LinearLayout
+<ScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingStart="8dp"
-    android:paddingEnd="8dp"
-    android:gravity="center_horizontal"
-    android:orientation="vertical"
-    android:id="@+id/root">
-
-</LinearLayout>
+    android:layout_height="match_parent">
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingStart="8dp"
+        android:paddingEnd="8dp"
+        android:gravity="center_horizontal"
+        android:orientation="vertical"
+        android:id="@+id/fragment_container" />
+</ScrollView>
